+++ commerce/delivery/src/main/java/ru/yandex/practicum/plus_smart_home_tech/delivery/service/DeliveryService.java
@@ -4,6 +4,7 @@ import jakarta.validation.Valid;
 import ru.yandex.practicum.plus_smart_home_tech.interaction_api.dto.delivery.DeliveryDto;
 import ru.yandex.practicum.plus_smart_home_tech.interaction_api.dto.order.OrderDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface DeliveryService {
@@ -15,5 +16,5 @@ public interface DeliveryService {
 
     void setFailed(UUID orderId);
 
-    Double calculateDeliveryCost(OrderDto orderDto);
+    BigDecimal calculateDeliveryCost(OrderDto orderDto);
 }
