+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/plus_smart_home_tech/interaction_api/dto/order/OrderDto.java
@@ -7,6 +7,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Map;
 import java.util.UUID;
 
@@ -31,7 +32,7 @@ public class OrderDto {
     private Double deliveryWeight;
     private Double deliveryVolume;
     private boolean fragile;
-    private Double totalPrice;
-    private Double deliveryPrice;
-    private Double productPrice;
+    private BigDecimal totalPrice;
+    private BigDecimal deliveryPrice;
+    private BigDecimal productPrice;
 }
